requirements.txt
typer
pydantic
pyyaml
pytest
lxml
beautifulsoup4
python-slugify
rich
docling
